pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tellodji"
version = "0.4.0"
description = "A Python library for controlling Tello drones using the 1.3 or 2.0 SDK"
authors = ["zahtec"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/zahtec/tellodji"
repository = "https://github.com/zahtec/tellodji"
documentation = "https://github.com/zahtec/tellodji/wiki"
keywords = [
    "tello",
    "drone",
    "drones",
    "sdk",
    "dji",
    "djitello",
    "interface",
    "video",
    "ffmpeg",
    "tellovideo",
    "tellodji",
    "djitello"
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    'Typing :: Typed'
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/zahtec/tellodji/issues"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.1.1"
